pyproject.toml
[project]
name = "text-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.1.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=0.3.1",
    "langgraph>=1.0.5",
    "matplotlib>=3.10.6",
    "networkx>=3.5",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
    "wikipedia>=1.4.0",
]
